Rename stimulus entrypoints from bootstrap to app to avoid confusion
