+++ src/Sylius/Bundle/AdminBundle/Resources/assets/app.js
