+++ src/Sylius/Bundle/ShopBundle/Resources/assets/app.js
