+++ src/Sylius/Bundle/ShopBundle/Resources/assets/entrypoint.js
@@ -9,7 +9,7 @@
 
 import './styles/main.scss';
 
-import './bootstrap';
+import './app';
 
 import './scripts/bootstrap';
